Avoid initial scrollTop set event
In https://github.com/elwayman02/ember-user-activity/pull/59 the initial
scrollTop is no longer read when setting up a subscriber. This avoids
reading the DOM while running JS and often forcing a layout invalidation.
However, it also means when the state is read an event and subscription
callback are fired. Since the value being set is actually a lazy initial
value and not a change, suppress that event.