Remove logo and multi-yield usage
That was in a fork, however it was never merged to upstream.
In fact that customisation doesn't give much to the look of the page.
It was adding the Ember logo to the hero section.
With the latest release of `ember-cli-addon-docs` it shows "Ember" label
in the `hero` section. So, it makes showing logo redundant.
It is safe just to remove the logo at all.